requirements.txt
google-genai
python-dotenv
pandas
numpy
requests
scikit-learn
fastapi
pydantic
streamlit
uvicorn
